+++ changelog.md
@@ -15,6 +15,7 @@
 * Added Gravity Defying enchantment
 * Added Elytra Nerf Feature
   * When using an elytra in non-End dimensions, the player falls faster
+* Forge recipes now award experience
 * Cleansed lapis now has greatly increased chance to drop
 * Expanded now works properly when looking down, and works as vein mining when on an Axe/Hoe. Also Expanded III now mines more blocks instead of being the same as lvl II
 * Buffed Feather Falling